backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
PyJWT==2.9.0
pydantic[email]==2.8.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
tenacity==8.5.0
